[FROM-ML] platform/x86: msi-wmi-platform: Add unlocked msi_wmi_platform_query
This driver requires to be able to handle transactions that perform
multiple WMI actions at a time. Therefore, it needs to be able to
lock the wmi_lock mutex for multiple operations.

Add msi_wmi_platform_query_unlocked() to allow the caller to
perform the WMI query without locking the wmi_lock mutex, by
renaming the existing function and adding a new one that only
locks the mutex.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>
Link: https://lore.kernel.org/all/20250511204427.327558-3-lkml@antheas.dev/
Signed-off-by: Matthew Schwartz <matthew.schwartz@linux.dev>
Signed-off-by: Derek J. Clark <derekjohn.clark@gmail.com>